Add files via upload
The main objective of IOT based noise pollution monitoring system is that the noise pollution is a rising issue these days.. Pollution is very harmful for those people who have any kind of internal diseases on this type of people pollution affect veryfastly. some are harmful for environment for certain level are good for human, animals, plants but beyond certain level these created problem for services to overcome these problem system is useful because of this we can analyse the noise pollution means how many pollution level in atmosphere in three different levels.We use Internet Of Things (IOT). In this we use thingspeak we can analyse previous data also using this platform in graphical form.